Moved event to dataset service. Refs #290430
